I  split the code into multiple files and separate our concerns for better readability and maintainability.
