+++ functions/index.js
@@ -1,211 +1,22 @@
 const functions = require("firebase-functions");
-const admin = require("firebase-admin");
-const app = require("express")();
-admin.initializeApp();
-
-const firebaseConfig = {
-  apiKey: "AIzaSyCGFyu6n87y7qzdRjSE_ULx21Gda3V2Kfo",
-  authDomain: "shitposter-1c203.firebaseapp.com",
-  databaseURL: "https://shitposter-1c203.firebaseio.com",
-  projectId: "shitposter-1c203",
-  storageBucket: "shitposter-1c203.appspot.com",
-  messagingSenderId: "361849568607",
-  appId: "1:361849568607:web:84d2db0dd19c65cb04caea",
-  measurementId: "G-X79Y3LNBMM",
-};
-
 const firebase = require("firebase");
+
+const firebaseConfig = require("./util/config");
 firebase.initializeApp(firebaseConfig);
 
-const db = firebase.firestore();
+const app = require("express")();
 
-app.get("/screams", (req, res) => {
-  db.collection("screams")
-    .orderBy("createdAt", "desc")
-    .get()
-    .then((data) => {
-      let screams = [];
-      data.forEach((doc) => {
-        screams.push({
-          screamId: doc.id,
-          body: doc.data().body,
-          userHandle: doc.data().userHandle,
-          createdAt: doc.data().createdAt,
-        });
-      });
-      return res.json(screams);
-    })
-    .catch((err) => console.error(err));
-});
+const FBAuth = require("./util/fbAuth");
 
-const FBAuth = (req, res, next) => {
-  let idToken;
-  if (
-    req.headers.authorization &&
-    req.headers.authorization.startsWith("Bearer ")
-  ) {
-    idToken = req.headers.authorization.split("Bearer ")[1];
-  } else {
-    console.log("no token found");
-    return res.status(403).json({ error: "Unauthorized" });
-  }
+const { getAllScreams, postOneScream } = require("./handlers/screams");
+const { signup, login } = require("./handlers/users");
 
-  admin
-    .auth()
-    .verifyIdToken(idToken)
-    .then((decodedToken) => {
-      req.user = decodedToken;
-      console.log(decodedToken);
-      return db
-        .collection("users")
-        .where("userId", "==", req.user.uid)
-        .limit(1)
-        .get();
-    })
-    .then((data) => {
-      req.user.handle = data.docs[0].data().handle;
-      return next();
-    })
-    .catch((err) => {
-      console.error("Error while verifying token", err);
-      return res.status(403).json(err);
-    });
-};
+// scream routes
+app.get("/screams", getAllScreams);
+app.post("/scream", FBAuth, postOneScream);
 
-app.post("/scream", FBAuth, (req, res) => {
-  const newScream = {
-    body: req.body.body,
-    userHandle: req.user.handle,
-    createdAt: new Date().toISOString(),
-  };
-  db.collection("screams")
-    .add(newScream)
-    .then((doc) => {
-      res.json({ message: `document ${doc.id} has been created successfully` });
-    })
-    .catch((err) => {
-      res.status(500).json({ error: "something went wrong on the server" });
-      console.error(err);
-    });
-});
-
-//validation functions
-const isEmpty = (string) => {
-  if (string.trim() === "") {
-    return true;
-  } else {
-    return false;
-  }
-};
-const isEmail = (email) => {
-  const emailRegEx = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  if (email.match(emailRegEx)) {
-    return true;
-  } else {
-    return false;
-  }
-};
-
-app.post("/signup", (req, res) => {
-  const newUser = {
-    email: req.body.email,
-    password: req.body.password,
-    confirmPassword: req.body.confirmPassword,
-    handle: req.body.handle,
-  };
-
-  let errors = {};
-
-  if (isEmpty(newUser.email)) {
-    errors.email = "Email must not be empty";
-  } else if (!isEmail(newUser.email)) {
-    errors.email = "Must be a vaild email";
-  }
-
-  if (isEmpty(newUser.password)) {
-    errors.password = "Password must not be empty";
-  }
-  if (newUser.password !== newUser.confirmPassword) {
-    errors.confirmPassword = "Passwords must match";
-  }
-  if (isEmpty(newUser.handle)) {
-    errors.handle = "handle must not be empty";
-  }
-
-  if (Object.keys(errors).length > 0) {
-    return res.status(400).json(errors);
-  }
-
-  // validate data
-  let token, userId;
-  db.doc(`/users/${newUser.handle}`)
-    .get()
-    .then((doc) => {
-      if (doc.exists) {
-        return res
-          .status(400)
-          .json({ handle: "this handle is already taken " });
-      } else {
-        return firebase
-          .auth()
-          .createUserWithEmailAndPassword(newUser.email, newUser.password);
-      }
-    })
-    .then((data) => {
-      userId = data.user.uid;
-      return data.user.getIdToken();
-    })
-    .then((idToken) => {
-      token = idToken;
-      const userCredentials = {
-        handle: newUser.handle,
-        email: newUser.email,
-        createdAt: new Date().toISOString(),
-        userId: userId,
-      };
-      return db.doc(`/users/${newUser.handle}`).set(userCredentials);
-    })
-    .then(() => {
-      return res.status(201).json({ token });
-    })
-    .catch((err) => {
-      console.log(err);
-      if (err.code === "auth/email-already-in-use") {
-        return res.status(400).json({ email: "Email is already taken" });
-      } else {
-        return res.status(500).json({ error: err.code });
-      }
-    });
-});
-
-app.post("/login", (req, res) => {
-  const user = {
-    email: req.body.email,
-    password: req.body.password,
-  };
-  let errors = {};
-  if (isEmpty(user.email)) return (errors.email = "Must not be empty");
-  if (isEmpty(user.password)) return (errors.password = "Must not be empty");
-  if (Object.keys(errors).length > 0) return res.status(400).json(errors);
-
-  firebase
-    .auth()
-    .signInWithEmailAndPassword(user.email, user.password)
-    .then((data) => {
-      return data.user.getIdToken();
-    })
-    .then((token) => {
-      return res.json({ token });
-    })
-    .catch((err) => {
-      console.log(err);
-      if (err.code === "auth/wrong-password") {
-        return res
-          .status(403)
-          .json({ general: "Wrong credentials, please try again" });
-      }
-      return res.status(500).json({ error: err.code });
-    });
-});
+// user routes
+app.post("/signup", signup);
+app.post("/login", login);
 
 exports.api = functions.https.onRequest(app);
