+++ functions/util/admin.js
@@ -1,0 +1,6 @@
+const firebase = require("firebase");
+const admin = require("firebase-admin");
+
+const db = firebase.firestore();
+
+module.exports = { admin, db };
