+++ functions/util/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  apiKey: "AIzaSyCGFyu6n87y7qzdRjSE_ULx21Gda3V2Kfo",
+  authDomain: "shitposter-1c203.firebaseapp.com",
+  databaseURL: "https://shitposter-1c203.firebaseio.com",
+  projectId: "shitposter-1c203",
+  storageBucket: "shitposter-1c203.appspot.com",
+  messagingSenderId: "361849568607",
+  appId: "1:361849568607:web:84d2db0dd19c65cb04caea",
+  measurementId: "G-X79Y3LNBMM",
+};
