+++ functions/util/validators.js
@@ -1,0 +1,51 @@
+//validation functions
+const isEmpty = (string) => {
+  if (string.trim() === "") {
+    return true;
+  } else {
+    return false;
+  }
+};
+const isEmail = (email) => {
+  const emailRegEx = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  if (email.match(emailRegEx)) {
+    return true;
+  } else {
+    return false;
+  }
+};
+
+exports.validateSignupData = (data) => {
+  let errors = {};
+
+  if (isEmpty(data.email)) {
+    errors.email = "Email must not be empty";
+  } else if (!isEmail(data.email)) {
+    errors.email = "Must be a vaild email";
+  }
+
+  if (isEmpty(data.password)) {
+    errors.password = "Password must not be empty";
+  }
+  if (data.password !== data.confirmPassword) {
+    errors.confirmPassword = "Passwords must match";
+  }
+  if (isEmpty(data.handle)) {
+    errors.handle = "handle must not be empty";
+  }
+
+  return {
+    errors,
+    valid: Object.keys(errors).length === 0 ? true : false,
+  };
+};
+
+exports.validateLoginData = (data) => {
+  let errors = {};
+  if (isEmpty(data.email)) return (errors.email = "Must not be empty");
+  if (isEmpty(data.password)) return (errors.password = "Must not be empty");
+  return {
+    errors,
+    valid: Object.keys(errors).length === 0 ? true : false,
+  };
+};
